c3_project/requirements.txt
appdirs==1.4.4
apturl==0.5.2
arabic-reshaper==2.1.3
asgiref==3.3.4
bcrypt==3.1.7
blinker==1.4
braintree==4.8.0
Brlapi==0.7.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2
django-activatable-model==2.0.0
django-braces==1.14.0
django-cors-headers==3.7.0
django-crispy-forms==1.7.0
django-environ==0.4.5
django-filter==2.4.0
django-forms-bootstrap==3.1.0
django-hosts==4.0
django-manager-utils==2.0.1
django-multiselectfield==0.1.12
django-oauth-toolkit==1.5.0
django-payments==0.13.0
django-payments-razorpay==1.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-query-builder==2.0.0
django-rest-framework-social-oauth2==1.1.0
django-tenant-schemas==1.10.0
djangorestframework==3.12.4
dnspython==1.16.0
drf-social-oauth2==1.0.9
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
fleming==0.6.0
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
gunicorn==19.7.1
html5lib==1.1
httplib2==0.14.0
idna==2.8
jwcrypto==0.8
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
ordered-set==4.0.2
paramiko==2.6.0
pbr==5.5.1
pexpect==4.6.0
phonenumberslite==8.12.20
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyDrive==1.3.1
PyGObject==3.36.0
PyJWT==2.0.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
PyPDF2==1.26.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-bidi==0.4.2
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2017.3
pyxdg==0.26
PyYAML==5.3.1
razorpay==1.2.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
SQLAlchemy==1.3.12
sqlparse==0.4.1
stripe==2.56.0
suds-jurko==0.6
systemd-python==234
testresources==2.0.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.4.3
virtualenvwrapper-win==1.2.6
wadllib==1.3.3
webencodings==0.5.1
xhtml2pdf==0.2.5
xkit==0.0.0
xmltodict==0.12.0
